+++ src/App.jsx
@@ -5,7 +5,7 @@ import { AppFooter, AppHeader } from './components';
 import { USER_ROUTE } from './pages/user/constants/routes-path/userRoute.constant';
 import LoginPage from './pages/login/LoginPage';
 import Register from './pages/register/RegisterPage';
-import PwdRestting from './pages/pwd-resetting/Password.resetting.component';
+import PwdRestting from './pages/pwd-resetting/PasswordResettingComponent';
 import { Chat } from './pages/chat/ChatMainPage';
 import AccommodationList from './pages/accommodation/AccommodationList.page';
 import Main from './pages/main/Main.page';
@@ -23,7 +23,7 @@ import AdminMain from './pages/admin/main/AdminMain.page';
 import ReservationManagementDetail from './pages/admin/reservation-detail/ReservationManagementDetail.component';
 import ReservationCancelList from './pages/admin/reservation-cancel/ReservationCancelList.page';
 import { AdminContactPage } from './pages/admin/contact/AdminContact.page';
-import ChatRoom from './pages/chat/chat-ui/Chat.room';
+import ChatRoom from './pages/chat/chat-ui/ChatRoomComponent';
 import LoginInterceptor from './pages/login/loginInterCepter/LoginInterceptor';
 import { ToastContainer } from 'react-toastify';
 import { loginStateStore } from './states/login/loginStore';
