+++ src/pages/chat/ChatMainPage.jsx
@@ -6,14 +6,14 @@ import {
 	PageContainer,
 	Select,
 } from '../../components';
-import { ChatTitle } from './chat-title/Chat.title.conponent';
-import './chatMainPage.scss';
+import { ChatTitle } from './chat-title/ChatTitleComponent';
+import './chatMainPageStyle.scss';
 import ChatStateStore from './chatStore';
 
 export function Chat() {
 	const { setMessage } = ChatStateStore();
 	return (
-		<PageContainer className={'chat-container'}>
+		<PageContainer className={'chat-container'}>G
 			<div className='chat-main-wrap'>
 				<ChatTitle
 					className={'chat-main-title'}
