+++ src/pages/chat/chat-title/ChatTitleComponent.jsx
@@ -1,4 +1,4 @@
-import './chat.title.conponent.scss';
+import './chatTitleStyle.scss';
 
 export function ChatTitle({ className, text }) {
 	return <p className={className}>{text}</p>;
