+++ src/pages/pwd-resetting/password-resetting/passwordResettingStyle.scss
