+++ src/pages/pwd-resetting/passwordResettingStyle.scss
