+++ src/pages/pwd-resetting/resetting-Email-auth/ResettingEmailAuthComponent.jsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
-import './resetting.email.auth.conponent.scss';
+import './resettingEmailAuthStyle.scss';
 import { ButtonPrimary } from '../../../components';
 import { useIsResetting } from '../../../states/pwdRestting/resetting.state';
-import { ResettingInput } from '../resetting-input/resetting.input.conponent';
-import { ResttingTitle } from '../resetting-title/Resetting.title.conponent';
+import { ResettingInput } from '../resetting-input/ResettingInputComponent';
+import { ResttingTitle } from '../resetting-title/ResettingTitleComponent';
 import { validateEmail } from '../../register/util/validate';
 
 export function EmailAuth() {
