+++ src/pages/pwd-resetting/resetting-Email-auth/resettingEmailAuthStyle.scss
