+++ src/pages/pwd-resetting/resetting-title/ResettingTitleComponent.jsx
@@ -1,5 +1,5 @@
 import { classNames } from '../../../utils';
-import './resetting.title.conponent.scss';
+import './resettingTitleStyle.scss';
 
 export function ResttingTitle({ className, text }) {
 	return <p className={classNames(className)}>{text}</p>;
