+++ src/pages/pwd-resetting/resetting-title/resettingTitleStyle.scss
