+++ src/pages/register/comple/CompleComponent.jsx
@@ -1,7 +1,10 @@
-import './comple.component.scss';
+import './compleStyle.scss';
 import { Link } from 'react-router-dom';
 import { ButtonPrimary } from '../../../components';
-import { RegisterInfostore, useRegisterStore } from '../../../states/register/RegisterStore';
+import {
+	RegisterInfostore,
+	useRegisterStore,
+} from '../../../states/register/RegisterStore';
 
 export default function RegisterComple() {
 	const { nickName } = RegisterInfostore();
