+++ src/pages/register/comple/compleStyle.scss
