+++ src/pages/register/email/emailStyle.scss
