+++ src/pages/register/nickName/NickName-sample/NickNameSampleComponent.jsx
