+++ src/pages/register/nickName/nickNameStyle.scss
