+++ src/pages/register/password/passwordStyle.scss
